feat: move accounts and storage related function
Closes #1178
