+++ packages/core/src/service/storage-management.ts
@@ -1,0 +1,37 @@
+import storage from '@onflow/flow-wallet-extension-shared/storage';
+
+import addressBookService from './addressBook';
+import coinListService from './coinList';
+import nftService from './nft';
+import transactionService from './transaction';
+import userInfoService from './user';
+import userWalletService from './userWallet';
+
+class StorageService {
+  clearNFT = () => {
+    nftService.clear();
+  };
+
+  clearNFTCollection = async () => {
+    await nftService.clearNFTCollection();
+  };
+
+  clearCoinList = async () => {
+    await coinListService.clear();
+  };
+
+  clearAllStorage = () => {
+    nftService.clear();
+    userInfoService.removeUserInfo();
+    coinListService.clear();
+    addressBookService.clear();
+    userWalletService.clear();
+    transactionService.clear();
+  };
+
+  clearLocalStorage = async () => {
+    await storage.clear();
+  };
+}
+
+export default new StorageService();
